+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 * Move internal helper functions to crawl4md.utils.check_helpers
 * Move markdown converter session pydantic models out of tests
 * Extract progress printing helper from tests support module
+* Extract markdown converter session loading and HTML normalization helpers
 
 ## [0.1.5] - 2026-05-03